Make spm_deface work with 4D data
The initial implementation of spm_deface failed to write 4D image data and the intention of direct and batch usage was inconsistent wrt to expected inputs. At the expense of not carrying all available NIFTI information, this code now treats 4D image data in input and output. Note, that each 3D volume in a 4D dataset will be defaced separately. An option like "other images" is not implemented in this tool.